feat: merge `test_*.py` TPC-H tests files
This PR merges the `test_*.py` files for the TPC-H tests into
`test_tpch.py`. The previous files only consisted of wrappers around the
generic test functions; the `test_X_on_Y.py` files even all consisted of
calls to `substrait_consumer_sql_test` with manually instantiated
producers and consumers. The new file uses fixtures to create producers
and consumers instead. Consequently, some combinations of producers and
consumers are now tested that previously weren't.
