+++ substrait_consumer/tests/integration/test_tpch.py
@@ -7,10 +7,9 @@ from pytest_snapshot.plugin import Snapshot
 from substrait_consumer.functional.common import (
     substrait_producer_sql_test,
     generate_snapshot_results,
+    substrait_consumer_sql_test,
 )
 from substrait_consumer.functional.utils import load_json
-from substrait_consumer.producers.duckdb_producer import DuckDBProducer
-from substrait_consumer.producers.isthmus_producer import IsthmusProducer
 
 CONFIG_DIR = Path(__file__).parent.parent
 TPCH_CONFIG_DIR = CONFIG_DIR / "integration" / "tpch"
@@ -18,6 +17,31 @@ TEST_CASE_PATHS = list(
     (path.relative_to(CONFIG_DIR),) for path in TPCH_CONFIG_DIR.rglob("*.json")
 )
 IDS = list((str(path[0]).removesuffix(".json") for path in TEST_CASE_PATHS))
+
+
+@pytest.mark.parametrize(["path"], TEST_CASE_PATHS, ids=IDS)
+@pytest.mark.usefixtures("prepare_tpch_parquet_data")
+def test_substrait_query(
+    path: Path,
+    snapshot: Snapshot,
+    db_con: duckdb.DuckDBPyConnection,
+    record_property,
+    producer,
+    consumer,
+) -> None:
+    test_case = load_json(CONFIG_DIR / path)
+    local_files = test_case["local_files"]
+    named_tables = test_case["named_tables"]
+    substrait_consumer_sql_test(
+        path,
+        snapshot,
+        record_property,
+        db_con,
+        local_files,
+        named_tables,
+        producer,
+        consumer,
+    )
 
 
 @pytest.mark.parametrize(["path"], TEST_CASE_PATHS, ids=IDS)
